feat: Add student status tracking
This commit introduces a `status` column to the `student` table in the database schema.

The `adminController` was failing when trying to load student data because it expected a `status` column that did not exist. This change adds the missing column, setting a default value of 'Inactive'.

This resolves the `java.sql.SQLException: Column 'status' not found` error.